backend/requirements.txt
fastapi
uvicorn[standard]
schedule
SQLAlchemy
pymysql
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
httpx
PyJWT
cryptography
itsdangerous